Modified XYLineAndShapeRenderer to allow the ability to define a custom logic that decides whether to draw a line between points.
For example:  If you're charting an angle in degress from 0 to 360 and when
it goes from 355 degrees to 361 degrees, you translated it back to 1 degrees 
instead of 361 degrees.  This example above will result to a chart that have a sharp
line connecting 355 degrees to 1 degrees.  If you don't wan't this, you can provide a 
drawLinesConditions logic that only draws line when the difference between current 
y-value and previous y-value is less than 180.